Use parsing context for constant value parsing
