+++ it/java8/src/main/java/online/sharedtype/it/java8/TempClass.java
@@ -3,7 +3,8 @@ package online.sharedtype.it.java8;
 import online.sharedtype.SharedType;
 import online.sharedtype.it.java8.other.OtherConstants;
 
+@SharedType(constantNamespaced = SharedType.OptionalBool.FALSE, includes = SharedType.ComponentType.CONSTANTS)
 public class TempClass extends IgnoredSuperClassB {
-    //    static final EnumGalaxy REFERENCED_VALUE = EnumGalaxy.MilkyWay;
+//        static final EnumGalaxy REFERENCED_VALUE = EnumGalaxy.MilkyWay;
 
 }
